Probable optimization.
By explicitly caching which states have Markovian transitions and which
transitions those are, we can hopefully find the transition set more
quickly than by exploring all the possible transitions.

Also, we are now using a special constructor for the transition set that
just trusts that the transitions are unique (we should probably verify
this by checking that the synchronization vectors don't allow duplicate
actions, but currently we trust the model).