Add support for timed transitions.
These are transitions in .exp/.aut models which are taken (independently
in each automaton) when the current real time is a multiple of the given
time.

Current implementation is only intended for use in MakeTraLab, and will
probably break horribly if an attempt is made to use it for simulation.