Add (strong) bisimulation reduction for automata.
A recent change to DFTCalc has caused the BE automata to have a
legitimate lasso pattern, which completely breaks the availability
simulation algorithm (as we never return to the original state).

We now reduce the automata modulo strong bisimulation, which removes
this lasso in most cases. We still have cases where it does not (namely,
BEs with a prob= attribute), in which case the steady-state simulator
will loop forever on the first trace.